requirements.txt
pandas
numpy
sklearn
tqdm
matplotlib
optuna
catboost
xgboost
lightgbm
keras
pyyaml
icecream
